refactor(components): move the stories that use components into components
Three stories reached into packages/components/src through relative paths.
Declaring that dependency where the files lived would have closed a cycle,
because components depends on tokens and on icons.

The stories now live in the package that sits above the one they document:
packages/components/stories/tokens/size.stories.tsx,
packages/components/stories/tokens/theme.stories.tsx and
packages/components/stories/icons/Icon.stories.tsx. Each keeps its title
(Tokens/Size, Tokens/Themes, Components/Icons/Icon), so Storybook navigation
and Chromatic snapshot identity do not move.

The moved files now read tokens through @launchpad-ui/tokens and icons
through @launchpad-ui/icons, both of which components already declares.
